fix(migration) miding duration column from and Interger to a float

+++ backend/migrations/20260405163432_track_schema.sql
@@ -4,7 +4,7 @@ CREATE TABLE tracks (
     id UUID PRIMARY KEY DEFAULT uuid_generate_v4(),
     title TEXT NOT NULL,
     artist TEXT,
-    duration_secs INTEGER NOT NULL CHECK (duration_secs >= 0),
+    duration_secs FLOAT NOT NULL CHECK (duration_secs >= 0),
     object_key TEXT NOT NULL,
     status track_status NOT NULL DEFAULT 'pending',
     created_at TIMESTAMPTZ NOT NULL DEFAULT NOW(),
